# src/main/java/ru/bmstu/iu5/doramadreams/exception/ErrorResponse.java
package ru.bmstu.iu5.doramadreams.exception;

import lombok.AllArgsConstructor;
import lombok.Data;

import java.time.LocalDateTime;

@Data
@AllArgsConstructor
public class ErrorResponse {
    private LocalDateTime timestamp;
    private Integer status;
    private String error;
    private String message;
    private String path;
}